chore: Split ArticleScreen Composable function to separate file
